# api_schemas/openapi.yaml
...
openapi: 3.0.3
info:
  title: AdaptiveMind AI API v1
  description: |
    Local-first AI orchestration platform providing chat, persona management, monitoring, and system configuration capabilities.
    
    ## Authentication
    All endpoints require authentication via the `X-API-Key` header. Set `ADAPTIVEMIND_DISABLE_AUTH=true` to bypass authentication locally.
    
    ## Endpoint Categories
    - **Core API** (`/api/v1/*`): Main AdaptiveMind functionality
    - **Management API** (`/api/v1/management/*`): System configuration and persona management
    - **OpenAI-Compatible** (`/v1/*`): Drop-in replacement for OpenAI API clients
  version: "1.0.0"
  contact:
    name: AdaptiveMind AI Support
    email: support@adaptivemind.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:8000
    description: Local development server
  - url: https://api.adaptivemind.ai
    description: Production server
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/chat:
    post:
      summary: Chat completion
      description: Generate a chat response using the specified persona and AI models
      tags: [Chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            example:
              messages:
                - role: user
                  content: "Hello, how are you?"
              persona: "generalist"
              temperature: 0.7
              max_tokens: 512
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              example:
                content: "I'm doing well, thank you for asking! How can I help you today?"
                model: "ollama"
                tokens: 25
                diagnostics: {}
        '400':
          description: Bad request (e.g., persona not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # Core API - Monitoring
  # ============================================================
...
  /api/v1/management/config/routing:
    put:
      summary: Update routing configuration
      description: Update routing configuration including allowed personas
      tags: [Management]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoutingConfigUpdateRequest'
      responses:
        '200':
          description: Updated routing configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutingConfigResponse'
        '400':
          description: Bad request (e.g., persona doesn't exist)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # Management API - Backends
  # ============================================================
...
  /api/v1/management/backends/{name}/test:
    post:
      summary: Test backend
      description: Test connectivity and availability of a specific backend
      tags: [Management]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Backend name to test
      responses:
        '200':
          description: Test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackendTestResponse'
              example:
                success: true
                latency_ms: 45.2
                error: null
        '404':
          description: Backend not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # Management API - Context
  # ============================================================
...
  /api/v1/management/config/context:
    put:
      summary: Update context configuration
      description: Update context pipeline configuration
      tags: [Management]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextConfigUpdateRequest'
      responses:
        '200':
          description: Updated context configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextConfigResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # Management API - Security
  # ============================================================
...
  /api/v1/management/personas:
    post:
      summary: Create persona
      description: Create a new persona configuration
      tags: [Management]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonaCreateRequest'
      responses:
        '200':
          description: Created persona
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaResponse'
        '400':
          description: Bad request (e.g., persona already exists)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/management/personas/{name}:
    put:
      summary: Update persona
      description: Update an existing persona configuration
      tags: [Management]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Persona name to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonaUpdateRequest'
      responses:
        '200':
          description: Updated persona
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaResponse'
        '404':
          description: Persona not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete persona
      description: Delete a persona configuration
      tags: [Management]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Persona name to delete
      responses:
        '200':
          description: Deletion confirmation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "Persona 'custom-persona' deleted successfully"
        '404':
          description: Persona not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # Management API - Config Save
  # ============================================================
...
  /v1/chat/completions:
    post:
      summary: OpenAI-compatible chat completions
      description: |
        Drop-in replacement for OpenAI's chat completions API.
        The `model` parameter is mapped to an AdaptiveMind persona if it exists, otherwise defaults to "generalist".
      tags: [OpenAI Compatible]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenAIChatRequest'
            example:
              model: "adaptivemind-default"
              messages:
                - role: user
                  content: "Hello!"
              temperature: 0.7
              max_tokens: 512
              stream: false
      responses:
        '200':
          description: Chat completion response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAIChatResponse'
              example:
                id: "chatcmpl-1702648475"
                object: "chat.completion"
                created: 1702648475
                model: "ollama"
                choices:
                  - index: 0
                    message:
                      role: "assistant"
                      content: "Hello! How can I help you today?"
                    finish_reason: "stop"
                usage:
                  prompt_tokens: 10
                  completion_tokens: 15
                  total_tokens: 25
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Error message
      required:
        - detail

    # ============================================================
    # Health
    # ============================================================
...
    ChatRequest:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
          description: Conversation messages
        persona:
          type: string
          default: "generalist"
          description: Persona to route the request
        temperature:
          type: number
          format: float
          default: 0.7
          minimum: 0.0
          maximum: 2.0
          description: Controls randomness in generation
        max_tokens:
          type: integer
          default: 512
          minimum: 32
          maximum: 4096
          description: Maximum tokens to generate
        metadata:
          type: object
          description: Optional metadata for the request
        external_context:
          type: array
          items:
            type: string
          description: Optional external context strings
      required:
        - messages
    ChatResponse:
      type: object
      properties:
        content:
          type: string
          description: Generated response text
        model:
          type: string
          description: Backend model used for generation
        tokens:
          type: integer
          description: Number of tokens generated
        diagnostics:
          type: object
          description: Backend-specific diagnostic information
      required:
        - content
        - model
        - tokens
        - diagnostics

    # ============================================================
    # Persona Models
    # ============================================================
...
    PersonaCreateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          pattern: '^[a-zA-Z0-9_-]+$'
          description: Unique persona name
        description:
          type: string
          minLength: 1
          maxLength: 200
          description: Persona description
        system_prompt:
          type: string
          minLength: 1
          description: System prompt for the persona
        max_context_window:
          type: integer
          default: 4096
          minimum: 512
          maximum: 32768
          description: Maximum context window size
        routing_hint:
          type: string
          default: "general"
          minLength: 1
          maxLength: 50
          description: Routing hint for backend selection
      required:
        - name
        - description
        - system_prompt
    PersonaUpdateRequest:
      type: object
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 200
          description: Updated description
        system_prompt:
          type: string
          minLength: 1
          description: Updated system prompt
        max_context_window:
          type: integer
          minimum: 512
          maximum: 32768
          description: Updated max context window
        routing_hint:
          type: string
          minLength: 1
          maxLength: 50
          description: Updated routing hint
    PersonaResponse:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        system_prompt:
          type: string
        max_context_window:
          type: integer
        routing_hint:
          type: string
        is_active:
          type: boolean
          description: Whether persona is in allowed_personas list
      required:
        - name
        - description
        - system_prompt
        - max_context_window
        - routing_hint
        - is_active

    # ============================================================
    # Monitoring Models
    # ============================================================
...
    RoutingConfigResponse:
      type: object
      properties:
        allowed_personas:
          type: array
          items:
            type: string
          description: List of personas allowed for routing
        enable_adaptive_routing:
          type: boolean
          default: true
          description: Whether adaptive routing is enabled
      required:
        - allowed_personas
        - enable_adaptive_routing
    RoutingConfigUpdateRequest:
      type: object
      properties:
        allowed_personas:
          type: array
          items:
            type: string
          description: Updated list of allowed personas
        enable_adaptive_routing:
          type: boolean
          description: Enable/disable adaptive routing
...
    BackendTestResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether backend is available
        latency_ms:
          type: number
          format: float
          description: Response time in milliseconds
        error:
          type: string
          nullable: true
          description: Error message if test failed
      required:
        - success
    ContextConfigResponse:
      type: object
      properties:
        extra_documents_dir:
          type: string
          nullable: true
          description: Directory for additional documents
        enable_semantic_chunking:
          type: boolean
          description: Whether semantic chunking is enabled
        max_combined_context_tokens:
          type: integer
          description: Maximum tokens for combined context
      required:
        - enable_semantic_chunking
        - max_combined_context_tokens
    ContextConfigUpdateRequest:
      type: object
      properties:
        extra_documents_dir:
          type: string
          description: Path to additional documents directory
        enable_semantic_chunking:
          type: boolean
          description: Enable/disable semantic chunking
        max_combined_context_tokens:
          type: integer
          minimum: 1024
          maximum: 65536
          description: Maximum tokens for context
...